refactor(coprocessor): move bootstrap_versioning to its own module
